requirements.txt
gym~=0.21.0
numpy~=1.23.4
torch~=1.13.0
scipy~=1.9.3
matplotlib~=3.6.2
mujoco~=2.3.0
wandb~=0.13.5
opencv-python~=4.6.0.66
setuptools~=65.5.0
tqdm~=4.64.1
pyyaml~=6.0